🧩 Modularization
