+++ entrypoints.ts
@@ -9,12 +9,12 @@ const sourceDir = './src/scripts';
  * @param dir Directory path to scan
  * @returns {Promise<string[]>} The entrypoints
  */
-async function getFiles(dir: string): Promise<string[]> {
+async function getFiles(dir: string, deep: boolean = false): Promise<string[]> {
   const dirents = await readdir(dir, { withFileTypes: true });
   const files = await Promise.all(
     dirents.map((dirent) => {
       const res = join(dir, dirent.name);
-      if (dirent.isDirectory()) {
+      if (dirent.isDirectory() && deep) {
         return getFiles(res);
       } else {
         return Promise.resolve(res);
