+++ public/content.css
@@ -113,3 +113,7 @@ table[aria-labelledby='folders-and-files']
   div {
   justify-content: flex-end;
 }
+
+.grs {
+  color: var(--fgColor-muted, var(--color-fg-muted));
+}
