--- src/constants/api.ts
@@ -1,1 +1,0 @@
-export const API: string = 'https://api.github.com/repos/';
