--- src/constants/selectors.ts
@@ -1,3 +1,0 @@
-export const NAV_ELEM_ID = 'github-repo-size';
-export const FILE_BUTTON = 'a.d-none.js-permalink-shortcut';
-// Should add more selectors here used in content.ts for clarity
