+++ src/scripts/internal/index.ts
@@ -1,0 +1,15 @@
+export { getPathObject } from './get-path-object';
+export { getRepoInfo } from './api';
+export { formatBytes } from './format';
+export { djb2, hashClass } from './crypto';
+export {
+  getNavButtons,
+  getAnchors,
+  getTotalSizeButton,
+  getSizeLabel,
+  getThead,
+  getTotalSizeSpan,
+} from './selectors';
+export { createSizeLabel, createTotalSizeButton } from './element-factory';
+export { updateDOM } from './dom-manipulation';
+export { getSize } from './get-size';
