+++ src/scripts/internal/types/index.ts
@@ -1,0 +1,6 @@
+export type {
+  PathObject,
+  GitHubTree,
+  GitHubTreeItem,
+  GRSUpdate,
+} from './types';
