+++ src/shared/index.ts
@@ -1,4 +1,3 @@
-export { API } from './api';
 export { GITHUB_TOKEN_KEY, OAUTH_TOKEN_KEY } from './storage-keys';
 export { default as storage } from './storage';
-export { NAV_ELEM_ID, FILE_BUTTON } from './selectors';
+export { getToken } from './get-token';
